Rename to tc_shortener.
