+++ tc_shortener/generators/__init__.py
